+++ CHANGES.rst
@@ -28,6 +28,8 @@ Bugfixes
 ^^^^^^^^
 
 - Fix incorrect order of session blocks inside timetable (:issue:`2999`)
+- Add missing email validation to contribution CSV import (:issue:`3568`,
+  thanks :user:`Kush22`)
 
 Internal Changes
 ^^^^^^^^^^^^^^^^